weather_app_flask/requirements.txt
Flask
Flask-CORS
httpx
google-generativeai
python-dotenv
